Fix few things

+++ code/core/trkCore.cc
@@ -648,7 +648,7 @@ void addInputsToLineSegmentTrackingPreLoad(std::vector<std::vector<float>> &out_
                                            std::vector<std::vector<float>> &out_eta_vec,
                                            std::vector<std::vector<float>> &out_etaErr_vec,
                                            std::vector<std::vector<float>> &out_phi_vec,
-                                           std::vector<std::vector<float>> &out_charge_vec,
+                                           std::vector<std::vector<int>> &out_charge_vec,
                                            std::vector<std::vector<int>> &out_superbin_vec,
                                            std::vector<std::vector<int8_t>> &out_pixelType_vec,
                                            std::vector<std::vector<short>> &out_isQuad_vec)
@@ -680,7 +680,7 @@ void addInputsToLineSegmentTrackingPreLoad(std::vector<std::vector<float>> &out_
     eta_vec.reserve(n_see);
     std::vector<float> phi_vec;
     phi_vec.reserve(n_see);
-    std::vector<float> charge_vec;
+    std::vector<int> charge_vec;
     charge_vec.reserve(n_see);
     std::vector<float> deltaPhi_vec;
     deltaPhi_vec.reserve(n_see);
@@ -915,7 +915,7 @@ float addInputsToEventPreLoad(SDL::Event* event,
                               std::vector<float> eta_vec,
                               std::vector<float> etaErr_vec,
                               std::vector<float> phi_vec,
-                              std::vector<float> charge_vec,
+                              std::vector<int> charge_vec,
                               std::vector<int> superbin_vec,
                               std::vector<int8_t> pixelType_vec,
                               std::vector<short> isQuad_vec)
